feat/videoClass:Arrumando arquivos e estilizações
