+++ src/components/VideoClass/Video/video.module.scss
@@ -1,0 +1,11 @@
+@import '../../../styles/sass/main.scss';
+
+iframe {
+    width: 973px;
+    height: 557px;
+    border-radius: 40px;
+    border: 5px solid #550084;
+    
+
+
+}
