+++ src/components/VideoClass/VideoClassTitle/videoClassTitle.module.scss
@@ -1,4 +1,4 @@
-@import '../../styles/sass/main.scss';
+@import '../../../styles/sass/main.scss';
 
 .container {
